+++ test/fixtures/beforestart/app.js
@@ -1,4 +1,3 @@
-'use strict';
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
   return new (P || (P = Promise))(function (resolve, reject) {
     function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
@@ -7,7 +6,9 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
     step((generator = generator.apply(thisArg, _arguments)).next());
   });
 };
-const sleep = require('ko-sleep');
+
+const { sleep } = require('../../utils');
+
 module.exports = function (app) {
   app.beforeStart(function() {
     app.beforeStartFunction = true;