+++ test/fixtures/boot/app.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const sleep = require('mz-modules/sleep');
+const { sleep } = require('../../utils');
 
 module.exports = class {
   constructor(app) {