+++ test/fixtures/egg/app/middleware/status.js
@@ -2,6 +2,7 @@
 
 module.exports = function() {
   return (ctx, next) => {
+    ctx.traceId = `trace:${Date.now()}`;
     if (ctx.path === '/status') {
       ctx.body = 'egg status';
       return;