+++ test/fixtures/timing/app.js
@@ -8,7 +8,14 @@ module.exports = app => {
 
   app.beforeStart(function* () {
     block();
-  })
+  });
+
+  app.beforeStart(function* () {
+    block();
+  }, 'mock Block');
+
+  const cb = app.readyCallback('mockReadyCallbackWithoutFunction');
+  setTimeout(cb, 1000);
 
   const directory = path.join(app.baseDir, 'app/proxy');
   app.loader.loadToContext(directory, 'proxy');