+++ test/loader/get_app_info.test.js
@@ -1,5 +1,3 @@
-'use strict';
-
 const path = require('path');
 const mm = require('mm');
 const assert = require('assert');
@@ -37,5 +35,11 @@ describe('test/loader/get_app_info.test.js', () => {
     mm(process.env, 'EGG_SERVER_ENV', 'default');
     app = utils.createApp('appinfo');
     assert(app.loader.appInfo.root === process.env.HOME);
+  });
+
+  it('should get scope when specified', () => {
+    mm(process.env, 'EGG_SERVER_SCOPE', 'en');
+    app = utils.createApp('appinfo');
+    assert(app.loader.appInfo.scope === 'en');
   });
 });